% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_corpus.R
\name{build_corpus}
\alias{build_corpus}
\title{Build a Corpus of Works from the Internet Archive}
\usage{
build_corpus(
  keywords,
  date_range = "1700 TO 1899",
  download_dir = "data-raw/corpus",
  max_results = 10000,
  chime = TRUE
)
}
\arguments{
\item{keywords}{A vector of keywords to search in the metadata of the Internet Archive's text collection}

\item{date_range}{The desired data range to search, specified in the format "yyyy TO yyyy"}

\item{download_dir}{The directory (relative to your working directory) to
which files from the Internet Archive will be downloaded.}

\item{max_results}{The maximum number of text results}

\item{chime}{Should the function chime on completion?}
}
\value{
A dataframe representing the corpus of downloaded texts
}
\description{
\code{build_corpus} downloads the OCR text versions of works found by searching the Internet Archive's metadata for the specified `keywords` over a given `date_range` (provided in the format "yyyy TO yyyy"), and it returns a dataframe that includes the Internet Archive's metadata about the retrieved works along with the location of the corresponding text files.
}
\details{
Details needed
}
\examples{
\dontrun{
 yf_corpus <- build_corpus(keywords = "yellow fever")
}

}
