% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_network_plot.R
\name{citation_network_plot}
\alias{citation_network_plot}
\title{citation_network_plot}
\usage{
citation_network_plot(citations, cited_only, ...)
}
\arguments{
\item{citations}{A dataframe representing classified citations such as that generated by \code{\link[classify_citations()]{classify_citations}}.}

\item{cited_only}{A dataframe representing the works that are only cited by---not citing---the other works in \code{citations}.  This need only be specified if multiple classifications are employed (e.g., "agree" and "disagree") and not all of the only-cited works fall into the first classification.  Otherwise, this information will be automatically generated from \code{citations}.}

\item{...}{Optional further parameters to pass along to \code{\link[GGally::ggnet2]{ggnet2}}.}
}
\value{
A ggplot object
}
\description{
\code{citation_network_plot} does stuff
}
\details{
The `citation_network_plot` function
}
\examples{

}
