% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_citing.R
\name{classify_citing}
\alias{classify_citing}
\title{Open Scanned Versions of Potentially Citing Works for Classification}
\usage{
classify_citing(df, save_dir = ".")
}
\arguments{
\item{df}{A dataframe representing downloaded texts (the "citing works") generated by `find_citing`}

\item{save_dir}{The directory to save the script that records the user's classifications.  This script will have the name "classify_[object specified as df].R".}
}
\value{
A dataframe
}
\description{
\code{classify_citing} opens the Internet Archive's scanned versions of each potentially citing work in a dataframe generated by `find_citing`, highlighting the potential citations
}
\details{
`classify_citing` facilitates the process of confirming that the potential citing works returned by `find_citing` actually do contain citations to the cited works.  
Such hand-checking is necessary because the `find_citing` function casts a broad net for citing works, returning all works that include the cited work's author that were published in or after the cited work's year of publication.
When `classify_citing` runs, it opens the Internet Archive's online, scanned version of each citing work (in order by author then date) in the default browser and displays all matches for a cited work. 
The user can then classify whether the potential citation is a false positive or a duplicate work, or rather an actual citation (perhaps with an additional qualification, such as "disagreeing").
As indicated by the prompt, after this determination has been made, the user should enter it and press the [return] key to proceed to the next potential citation.

Although classifications can be entered as text (e.g., "false positive", "disagreeing"), a coding scheme such as the following can make the process easier:

 -1 duplicate source or similar later edition
 0  false positive
 1  agreeing with cited work or citing work as authority
 2  disagreeing with cited work 

The entered classifications are returned as a new variable named `classification` and are saved as an `R` script named "classify_[object specified as df].R" in the directory specified in the `save_dir` argument.
}
\examples{
\dontrun{
cites_rush <- classify_citing(mentions_rush) \%>\% 
     filter(classification > 0)
}


}
\seealso{
\code{\link{find_citing}}
}
