% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_all.R
\name{match_all}
\alias{match_all}
\title{Generate a Regular Expression for Finding Multiple Patterns in a Text}
\usage{
match_all(pattern1, pattern2 = "")
}
\arguments{
\item{pattern1, pattern2}{Two character strings each containing a word or a \link[regular expression]{stringi-search-regex} to be found in the same text.  Alternately, pattern1 can be a character vector.}

\item{max_distance}{The maximum number of words intervening between the two specified patterns}
}
\value{
A regular expression pattern
}
\description{
Generate a Regular Expression for Finding Multiple Patterns in a Text
}
\details{
\code{match_all} is a convenience function that creates a regular expression pattern for finding if two words (or other regular expression patterns) occur within a text
}
\examples{
colors <- "red|blue|yellow"
sentences <- c("The car parked by their house was red",
               "Her yellow cat enjoys the sun on the windowsill",
               "The blue city bus passed right by the stopped truck")
stringr::str_detect(sentences, match_all("yellow", "car"))
stringr::str_detect(sentences, match_all(colors, "car"))
stringr::str_detect(sentences, match_all(colors, "car|bus"))
stringr::str_detect(sentences, match_all(c("bus", "truck", "city")))

}
