% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_near.R
\name{match_near}
\alias{match_near}
\title{Generate a Regular Expression for Finding Two Patterns Occurring Near Each Other}
\usage{
match_near(pattern1, pattern2, max_distance)
}
\arguments{
\item{pattern1, pattern2}{Two character strings each containing a word or a \link[regular expression]{stringi-search-regex} to be found near each other.}

\item{max_distance}{The maximum number of words intervening between the two specified patterns}
}
\value{
A regular expression pattern
}
\description{
Generate a Regular Expression for Finding Two Patterns Occurring Near Each Other
}
\details{
\code{match_near} is a convenience function that creates a regular expression pattern for finding if two words (or other regular expression patterns) occur within a specified number of words of each other
}
\examples{
colors <- "red|blue|yellow"
sentences <- c("The car parked by their house was red",
               "The yellow cat enjoys the sun atop the car's roof",
               "The blue city bus passed right by the stopped car")
stringr::str_detect(sentences, match_near(colors, "car", 5))
stringr::str_detect(sentences, match_near(colors, "car|bus", 5))
stringr::str_detect(sentences, match_near(colors, "car", 10))


}
