% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_vector.R
\name{match_vector}
\alias{match_vector}
\title{Generate a Regular Expression for Finding Any of a Vector of Patterns}
\usage{
match_vector(pattern)
}
\arguments{
\item{pattern}{A character vector containing a series of words or \link[regular expression]{stringi-search-regex}s, any one of which is sufficient for a match}
}
\value{
A regular expression pattern
}
\description{
Generate a Regular Expression for Finding Any of a Vector of Patterns
}
\details{
\code{match_vector} is a convenience function that creates a regular expression pattern for finding any of a vector of words (or other regular expression patterns)
}
\examples{
colors <- c("red", "blue", "yellow")
sentences <- c("The car parked by their house was red",
               "The yellow cat enjoys the sun atop the car's roof",
               "The blue city bus passed right by the stopped car")
stringr::str_detect(sentences, match_vector(colors))


}
