% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_corpus.R
\name{subset_corpus}
\alias{subset_corpus}
\title{Return Works of a Corpus that Include Matching Text}
\usage{
subset_corpus(corpus, pattern)
}
\arguments{
\item{pattern}{A character string containing a series of words or a \link[regular expression]{stringi-search-regex}.}

\item{corpus_df}{A dataframe representing a corpus of downloaded texts generated by \code{\link[build_corpus()]{build_corpus}}.}
}
\value{
A dataframe
}
\description{
\code{subset_corpus} returns the works in a corpus whose texts include all of the words (or match the \link[regular expression]{stringi-search-regex}) specified in the `pattern` argument
}
\details{
\code{subset_corpus} returns the works in a corpus whose texts include all of the words (or match the \link[regular expression]{stringi-search-regex}) specified in the `pattern` argument.  It can therefore be used to identify a subset of the corpus that is of particular interest.
}
\examples{

}
