% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_network_plot.R
\name{citation_network_plot}
\alias{citation_network_plot}
\title{citation_network_plot}
\usage{
citation_network_plot(citations, cited_only, color_plot = TRUE,
  custom_plot = FALSE, arrow_gap = 0.015)
}
\arguments{
\item{citations}{A dataframe representing classified citations such as that generated by \code{\link[classify_citations()]{classify_citations}}.}

\item{cited_only}{A dataframe representing the works that are only cited by---not citing---the other works in \code{citations}.  This need only be specified if multiple classifications are employed (e.g., "agree" and "disagree") and not all of the only-cited works fall into the first classification.  Otherwise, this information will be automatically generated from \code{citations}.}

\item{color_plot}{Should multiple classifications be denoted by different colors?  If not, they will indicated by different shapes.}

\item{custom_plot}{Should the function return a bare plot suitable for customization using \code{\link[ggnetwork]{ggnetwork}} functions?}

\item{arrow_gap}{A parameter that will shorten the arrows that link works in order to avoid overplotting the dots that represent the works; larger values may be desirable if \code{custom_plot} is set to TRUE and the size of the dots will be increased using \code{\link[ggnetwork]{ggnetwork}}'s \code{\link[geom_nodes]{geom_nodes}} function.}
}
\value{
A ggplot object
}
\description{
\code{citation_network_plot} generates plots of citation networks, depicting works as dots ("nodes", in network parlance) and citations as arrows ("edges") from cited works to citing works.
}
\details{
The `citation_network_plot` function
}
\examples{

}
