% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_citing.R
\name{find_citing}
\alias{find_citing}
\title{find_citing}
\usage{
find_citing(corpus_df, df, verbose = TRUE)
}
\arguments{
\item{corpus_df}{A dataframe representing a corpus of downloaded texts generated by \code{\link[build_corpus()]{build_corpus}}.}

\item{df}{A dataframe representing downloaded texts (the "cited works") whose citation histories are sought; see details.}

\item{verbose}{Should the function provide information about its progress?}
}
\value{
A dataframe
}
\description{
\code{find_citing} does stuff
}
\details{
The `find_citing` function casts a broad net for citing works, returning all works that include the cited work's author's last name that were published in or after the cited work's year of publication.
}
\examples{
rush1794 <- yf_corpus \%>\% 
    filter(id = "101283166.nlm.nih.gov")

mentions_rush <- yf_corpus \%>\% 
    find_citing(rush1794)

}
